Improvements

+++ source/admin_page/admin_page.php
@@ -63,7 +63,8 @@
 </div>
 
 <?php
-	  $sql = "select * from Checkin where loc_id = '".$admin_location_id."';";
+	  //$sql = "select * from Checkin where loc_id = '".$admin_location_id."';";
+	$sql = "select * from (select * from Checkin where loc_id = '".$admin_location_id."' order by time asc) as checkin left join (select checkin_id, count(*) as num_of_likes from checkin_likes group by checkin_id) as checkin_likes on checkin.id = checkin_likes.checkin_id";
 	  $result = $conn->query($sql);
 	  if($result->num_rows > 0){
 		while($row = $result->fetch_assoc() ){
@@ -76,12 +77,17 @@
 		  }
 			  echo "<div class= "."rectangle"." >";
 				echo "<div class= "."column left"." style= "."background-color:#aaa;".">";
-				echo "<img src="."images/elon.png".">";
+				if (file_exists("../images/profile".$row['user_id'].".png")) {
+					echo "<img src=\"../images/profile".$row['user_id'].".png\" width=\"200\">";
+				}
+				else {
+					echo "<img src="."images/elon.png".">";
+				}
 				echo "</div>";
 				echo "<div class="."column right"." style="."background-color:#black;".">";
 				echo "<a href="."../check_in_comment/check_in_comment.php?var=".$row['id']."&var2=".$row['user_id'].""."><font size="."5".">".$checker_name." Has checked-in: ".$admin_location_name."</font></a>
 			&emsp;<p>".$checker_name."'s comment: ".$row['text']."</p><br>
-			<p>".$row['time'].".&emsp;&emsp;&emsp;report button&emsp;&emsp; number of like: ".$row['num_of_likes']."</p>";
+			<p>".$row['time'].".&emsp;&emsp;&emsp;Number of likes: ".(int)$row['num_of_likes']."</p>";
 				echo "</div>";
 			  echo "</div>";
 			  echo "<hr class=style1  width=60%> ";
