+++ source/myProfile/myProfile.php
@@ -21,7 +21,7 @@ $_SESSION['id'] = $id;
   <a href="../friends/friends.php?var=<?php echo $id ?>">Friends</a>
   <a class="active" href="../myProfile/myProfile.php?var=<?php echo $id ?>">myProfile</a>
   <a href="../settings/settings.php?var=<?php echo $id ?>">Settings</a>
-  <a href="../index.php">Logout</a>
+  <a href="../logout.php">Logout</a>
   <div class="search-container">
     <form action="../search_location/search_location.php">
       <input type="text" placeholder="Search.." name="search">
@@ -46,39 +46,61 @@ $_SESSION['id'] = $id;
     }
   }
 
-  $sql = "select bio from User where user_id = '".$user_id."' ;";
+  $sql = "select * from (select * from User where user_id = ".$user_id.") as user join (select user_id, count(*) as num_of_checkins from checkin group by user_id) as checkin_num on user.user_id = checkin_num.user_id";
+  $is_user = false;
   $result = $conn->query($sql);
   if($result->num_rows > 0){
     while($row = $result->fetch_assoc() ){
       $bio = $row['bio'];
-    }
-  }
-
-  $sql = "select birth_date from User where user_id = '".$user_id."' ;";
-  $result = $conn->query($sql);
-  if($result->num_rows > 0){
-    while($row = $result->fetch_assoc() ){
       $birth_date = $row['birth_date'];
+      $num_of_checkins = $row['num_of_checkins'];
+		$is_user = true;
     }
   }
+	$loc_man;
+	if(!$is_user) {
+		$sql = "SELECT * FROM LocationAdmin WHERE user_id = ".$user_id;
+		$result = $conn->query($sql);
+		if($result->num_rows > 0){
+			while($row = $result->fetch_assoc() ){
+				$loc_man = $row;
+			}
+		}
+	}
 
 ?>
 
 <div class="rectangle">
   <div class="column left" style="background-color:#aaa;">
+<?php
+	if (file_exists("../images/profile".$user_id.".png")) {
+		echo "<img src=\"../images/profile".$user_id.".png\" width=\"200\">";
+	}
+	else {
+?>
     <img src="images/elon.png">
+<?php
+	}
+?>
   </div>
   <div class="column right" style="background-color:#bbb;">
-    <h2><?php echo $bio; ?></h2>
+    <h2><?php echo $is_user ? $bio : $loc_man['role']; ?></h2>
+<?php
+	if($is_user) {
+?>
     <p>Birth date: <?php echo $birth_date; ?></p>
-    <p>number of checkins: <?php echo 'database content'; ?></p>
+    <p>number of checkins: <?php echo $num_of_checkins; ?></p>
+<?php
+	}
+?>
   </div>
 </div>
 
 <hr class=style1  width="60%">
 
 <?php
-  $sql = "select * from Checkin where user_id = '".$user_id."' order by time asc;";
+  //$sql = "select * from Checkin where user_id = '".$user_id."' order by time asc;";
+  $sql = "select * from (select * from Checkin where user_id = ".$user_id.") as checkin left join (select checkin_id, count(*) as num_of_likes from checkin_likes group by checkin_id) as checkin_likes on checkin.id = checkin_likes.checkin_id";
   $result = $conn->query($sql);
   if($result->num_rows > 0){
     while($row = $result->fetch_assoc() ){
@@ -92,12 +114,17 @@ $_SESSION['id'] = $id;
 
       echo "<div class= "."rectangle"." >";
         echo "<div class= "."column left"." style= "."background-color:#aaa;".">";
-        echo "<img src="."images/elon.png".">";
+		if (file_exists("../images/profile".$user_id.".png")) {
+			echo "<img src=\"../images/profile".$user_id.".png\" width=\"200\">";
+		}
+		else {
+			echo "<img src="."images/elon.png".">";
+		}
         echo "</div>";
         echo "<div class="."column right"." style="."background-color:#black;".">";
         echo "<a href="."../check_in_comment/check_in_comment.php?var=".$row['id']."&var2=".$row['user_id'].""."><font size="."5".">".$id." Has checked-in: ".$location_name."</font></a>
         &emsp;<p>".$id."'s comment: ".$row['text']."</p><br>
-        <p>".$row['time'].".&emsp;&emsp;&emsp;report button&emsp;&emsp; number of like: ".$row['num_of_likes']."</p>";
+        <p>".$row['time'].".&emsp;&emsp;&emsp;Number of likes: ".(int)$row['num_of_likes']."</p>";
         echo "</div>";
       echo "</div>";
       echo "<hr class=style1  width=60%> ";
